Reorganizing project structure
Reorganizing project structure and namespaces to fit (API + DB) for separation of concerns
