+++ api/controllers/HomeController.php
@@ -1,15 +1,15 @@
 <?php
 namespace API\Controllers;
 
-use \Database\Models\User;
+//use \DB\Models\User;
 
 class HomeController 
 {
   public function Index()
   {
-    $user = new User;
+    //$user = new User;
     
-    print_r($user->Find(2));
+    //print_r($user->Find(2));
   }
   public function Test($id, $data) {
     echo $id . " and " . $data;
