+++ api/controllers/UserController.php
@@ -2,7 +2,6 @@
 namespace API\Controllers;
 
 use \API\Core\Authenticator;
-use \Database\Models\User;
 
 class UserController extends Authenticator
 {
