+++ api/core/Helpers.php
