+++ config.php
@@ -1,8 +1,5 @@
 <?php
-// holds all global configuration variables
-return
-[
-  /* Database definitions */
+return [
   "DB_NAME" => "jobsportaldb",
   "DB_HOST" => "localhost",
   "DB_PORT" => 3306,
@@ -14,4 +11,5 @@ return
     PDO::ATTR_DEFAULT_FETCH_MODE => PDO::FETCH_ASSOC,
     PDO::ATTR_EMULATE_PREPARES   => false,
   ]
+
 ];
